fft2d fix, utils folder created

+++ src/ffte.c
@@ -6,7 +6,7 @@
 #include "ffte/subfuncs.h"
 
 // FFT
-void ffte(double* x_real, double* x_imag, size_t N, unsigned char only_real_input)
+void ffte(double* x_real, double* x_imag, unsigned int N, unsigned char only_real_input)
 {
 	if(N<2)
 		return;
@@ -17,7 +17,7 @@ void ffte(double* x_real, double* x_imag, size_t N, unsigned char only_real_inpu
 }
 
 // IFFT
-void iffte(double* x_real, double* x_imag, size_t N, unsigned char only_real_input)
+void iffte(double* x_real, double* x_imag, unsigned int N, unsigned char only_real_input)
 {
 	if(N<2)
 		return;
